Support onPreKeyDown
You can now pass in a function to onPreKeyDown, which gets called before
the default `KeyboardHandlerMixin` gets called. This is useful if you
want to implement something like an app-wide ⌘  or Control feature like
search or global undo.

Additionally, I updated `KeyboardHandlerMixin` to treat the ⌘  key the
same way that it treats the control key.